Correct is{First,Last}InSection to use .getDate instead of .getDay. (Fixes #15)
The error would cause the date header in the trip history to not be shown if
two trips were in the same year and month but the same day of the week.  For
example, two trips on 2016-03-16 and 2016-03-23 would only ever show one of
the headers, because they are on the same day of the week (Wednesday) in the
same month and year.