Fix import and export functionality.
Export all pushes the file size up by a lot. So old code doesn\t work with this
file size. This fixes several issues:

* Writing/reading files/sharing has to be done in async tasks with at most
a weak reference to fragment (strictmode and everything)
* file is too large to fit in a parcel atached to intent. So it needs to be
saved and then passed as an URI.
* Missing close() call resulted in 0-size files when saving to drive

Paths for pre-KitKat were left alone as it's less of an issue.