Fix potential crash due to reusing view of incorrect type.
If a card has many subscriptions, it can happen that view are reused (convertView is non-null).
If a BalanceView is reused as a SubscriptionView (or vice versa), this leads to a crash of the
application as findById returns null for sub-views.

When a card does not have many subscriptions, the bug is not triggered here, as the views are
not reused.

I used getTag()/setTag() to mark the views and check whether they are the correct type. An easier
fix would be to always discard "convertView". You could also check for a unique view id with
findById(), but IMHO this is more error-prone.