kotlin-mpp: Step 1: Convert tests to mpp
This is a move to kotlin-multiplatform.
Kotlin-mpp allows compilation for JS, JVM and native. In my tests in
addition to what we currently have I also run tests in desktop JVM,
have JVM cli and ChromeOS app

This commit converts to multiplatform tests so that the same tests can be
run on all the platforms. OBviously we still can add platform-specific
tests but currently all tests except currency formatting apply to all the platforms