Support desfire cards with locked directory list
MagnaCarta locks down the directory list. It doesn't prevent us from iterating
through files as there are only 256 possible file IDs and only 32 that are
ever used in any parser. However we can't read file settings in this case
so instead we have to try all 3 possible read command and then save which one
worked for recreating the parsed files