Use FormattedString throughout Trip.
Currently we don't track which parts of route, agency and station names are
in which language and then rendering fragments takes wild guesses at what
language is used where.
This started as a quick fix to fix spans on the string
"Valid from <Station> to <Station>" but has quickly grown out of it and
exposed a systematic problem. So here we have this rather large commit