Add functions to localize preserving formatting
Currently in the sentence "Valid from ... to ..." we drop all formatting
including localespan.
Android doesn't have a convenient String.format equivalent for formatted string
iOS doesn't have String.format that works for us as %s assumes a string in local
encoding. %S would work but changing format identifier is as tedious as
formatting ourselves.
So implement FormattedString.format that we can use in those cases and
localizeFormatted to preserve formatting of the arguments