subscription: Add number of zones to the string.
Some languages e.g. Russian have different handling for plurals when
the numeral is present and when it's not. E.g. in Russian
For 1: "1 Zona" or simply "Zona"
For 21: "21 Zona" but "Zony" if number is not explicitly specified
As you can see in the presence of numeral both 1 and 21 use the same form but
not in its absence.
Unfortunately Android uses form "one" for both 1 and 21. So it supports only
strings with explicit numeral.
So add the explicit numeral