Support reading Mifare Plus
Since Mifare plus is essentially a Mifare Classic with a longer key.
It's even possible for some sectors to be in classic and some other in plus mode
This patch treats it as a variant of mifare classic
The transit factory list is split to avoid needlessly iterating through wrong
factories but one factory may be in both lists.