iOS: Remove metrodroid registration as "Open with" target
I was under mistaken impression that document picker didn't exist on iOS.
Hence I registered Metrodroid as a reader for zip and json. As
document picker does exist and now we use it and we're not a real viewer for
either of formats, it is probably better not to pollute "share with" selection
and that decreases probability that someone tries to open unrelated zip or
unrelated json with us leading to unexpected results.