Fix-up MainActivity layout regression.
There was a linter warning advising about a useless RelativeLayout and/or
LinearLayout.  Before 48546934721182565dd4726545f640c0481d6ebc, this was
Linear -> Relative -> Linear.  48546934721182565dd4726545f640c0481d6ebc
changed this to just Linear, but it moves all the UI elements to the top of
the screen.

This partially reverts the commit, such that it only has Relative -> Linear,
add disables the linter warning.  That way, all the UI elements are centred
again.