ios: Work around #613 (cannot read multi-system FeliCa cards)
This adds a new parameter to `FelicaReader.dumpTag`, `onlyFirst`.  This
causes any system codes after index=0 to be skipped, and an empty
`FelicaService` is inserted in place (and the event is logged).

This will probably only read the Octopus part of Hu Tong Xing (互通行) --
unfortunately this can't be fixed without having the underlying iOS bug
fixed.