Clipper: Fix vehicle numbers for LRV4 Muni Vehicles
Historically, Muni just stored a 4-digit (decimal) vehicle number as part of a
Trip record on the Clipper Card. Muni rolled out a line of new LRV4 Light Rail
cars a while ago with newer Clipper Card readers. These readers also store
which car you are in (A vs B) for multi-car trains, in addition to the 4-digit
vehicle number, but as an extra decimal digit at the end, mapping 0d1 => A, 0d2
=> B, etc. For example, vehicle number 2046A is stored as 0d20461 (0x3fed),
2046B is stored as 0d20462 (0x4fee), etc.

This change modifies the ClipperTrip vehicleID method to take this into account
and adds some tests for it along the way.

The Trip rows come from an actual Clipper card that I've been using.