Remove dependency on kotlinx.io
Last release dates 2019 and it's no longer compatible with recent
kotlin/native. As we need very little from it, just use small own wrappers

Some of the changes here are not strictly necessarry, just for the readability
and code simplification

@UseExperimental(ExperimentalStdlibApi::class) is needed because
encodeToByteArray was marked as Experimental. It's not anymore and we can
remove it after updating kotlin