ImmutableByteArray.sliceOffLenSafe: return empty ImmutableByteArray when len = 0
This fixes an issue in the BER-TLV parser where if the last tag was
zero-length, it would be missed.  The blank Mobib card actually contains one
of these in the `TICKETING_ENVIRONMENT` FCI (0x20), so need to update the test there.

My device with Apple Pay contained one such tag as well.

This adds a test for the change in ImmutableByteArray's behaviour, and some
isolated BER-TLV test data.

This exposed an issue in the FeliCa parser which depended on the existing
ImmutableByteArray behaviour (using `Int.rangeOf` rather than `Int.until`).