MetrodroidZipTest: fix timestamps on API <= 25
The existing ZIP file had some timestamps that were on an odd second (files
7 and 8), which isn't representable using the original DOS timestamps.
Also, it looks like the archive was created with a timezone other than
`America/Los_Angeles` (used by the test).

This fudges the timestamp forward 1 second, and also compresses the archive
with `TZ=America/Los_Angeles`.  This allows the tests to pass regardless of
whether the DOS or UNIX timestamps are used.